split up file into multiple
