+++ packages/core/src/generators/solid/state/index.ts
@@ -1,0 +1,1 @@
+export { getState } from './state';
